+++ helper.h
@@ -11,7 +11,7 @@ Roots linear_solve(const Scalars scalars);
 Roots quadratic_solve(const Scalars scalars);
 
 /*atof which handles ','*/
-float almost_my_atof(const char s[]);
+double almost_my_atof(const char s[]);
 
 /*checks if program is run with correct arguments*/
 Scalars parse_arguments(int argc, char *argv[]);
@@ -19,8 +19,8 @@ Scalars parse_arguments(int argc, char *argv[]);
 /*checks if argument is a valid number*/
 int is_argument_valid(const char* const arg);
 
-/*compares two float with EPS precision*/
-int float_equals(const float a, const float b, const float eps);
+/*compares two double with EPS precision*/
+int float_equals(const double a, const double b, const double eps);
 
 /*prints "documentaion"*/
 void print_doc();
